// backend/src/helper/timetableHelper.d.ts
export declare const normalizeTimetable: (timetableRows: any[], requestedDay: string) => {
    id: any;
    period: any;
    time: string;
    startTime: any;
    endTime: any;
    isActive: boolean;
    isBreak: any;
    breakLabel: any;
    room: any;
    color: any;
    subject: any;
    professor: string | null;
}[];
//# sourceMappingURL=timetableHelper.d.ts.map